backend/requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
pydantic[email]
python-multipart==0.0.6
mangum==0.10.0